int to size_t

+++ libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: fras <fras@student.codam.nl>                 +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/10/17 17:25:39 by fras          #+#    #+#                 */
-/*   Updated: 2022/10/18 09:41:36 by fras          ########   odam.nl         */
+/*   Updated: 2022/10/25 20:49:23 by fras          ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
